+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.8.0] - 2021-11-01
+
 ### Added
 - Button "Copy" to copy a code snippet to the clipboard (#27)
 
@@ -98,7 +100,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Syntax highlighting with Prism
 
 
-[Unreleased]: https://github.com/brotkrueml/codehighlight/compare/v2.7.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/codehighlight/compare/v2.8.0...HEAD
+[2.8.0]: https://github.com/brotkrueml/codehighlight/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/brotkrueml/codehighlight/compare/v2.6.1...v2.7.0
 [2.6.1]: https://github.com/brotkrueml/codehighlight/compare/v2.6.0...v2.6.1
 [2.6.0]: https://github.com/brotkrueml/codehighlight/compare/v2.5.0...v2.6.0